backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0

# Speech-to-Text (Groq Whisper chunking for files > 24MB)
# pydub>=0.25.1

# Google Gemini
google-generativeai>=0.8.0

# Telegram Bot
python-telegram-bot>=21.0

# Utilities
python-dotenv==1.0.0
pydantic>=2.7.4
pydantic-settings>=2.10.1
aiofiles==23.2.1
orjson>=3.9.0
fpdf2>=2.7.8
prometheus-client>=0.20.0
redis[hiredis]>=6.2.0,<8.0.0
boto3>=1.34.0

# HTTP Client
httpx>=0.27.0
pyTelegramBotAPI

# Authentication
PyJWT>=2.8.0
bcrypt>=4.1.0